requirements.txt
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
gradio==4.31.2